Fix crash when stopping the processing.
There was couple of problems in handling the internal state - it was too
optimistic about timing. And it didn't handle correctly the stopping states.
Fixed this by adding new attribute for tracking if there is real activity
going on. And allow the stopping to complete only after the activity has really
stopped.